Split up Angle impl + tests into separate module
